python/requirements.txt
fastapi
uvicorn
python-dateutil
requests